Join a Workstation to the Domain
When you're setting up your lab, you probably want to join a computer to your domain. This script accomplishes this. It will ask for a domain admin UN/PW when you run this. 

Pre-reqs is that your device must have a line of site to your domain controller, so you'd want to ensure the endpoint's DNS on the Network controller is set to the IP address of the Domain Controller. That will allow you to authenticate w/ Domain Admin privs and join it to the domain.