Add EntraID UPN Suffix is On-Prem AD Accounts
Determine your cloud UPN suffix and add it to all of the accounts in your local domain. This will allow your users to access resources in your tenant when you sync them. If you have a .com domain name, you could also add that provided you set this up as a trusted domain in EntraID.